Fix PPM support when using 8 channels on STM32F10x.
This was broken when 12 channel support was added in combination with
the STM32F3 support.

+++ src/drivers/timer_common.h
@@ -10,7 +10,7 @@
 typedef uint32_t captureCompare_t;
 #endif
 #ifdef STM32F10X_MD
-typedef uint32_t captureCompare_t;
+typedef uint16_t captureCompare_t;
 #endif
 
 typedef void timerCCCallbackPtr(uint8_t port, captureCompare_t capture);
